+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ## 2.0.1
 
+- Added correct `withMetadata` typing to `AssetTransfersResponse`.
+
 ### Minor Changes
 - Added an optional `url` setting to `AlchemySettings` to allow specifying a custom hardcoded URL to send all requests to.
 - Fixed a bug where the SDK would incorrectly include a gzip header in browser environments