+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Minor Changes
 - Added support for the `pageSize` parameter in `GetNftsForOwnerOptions` to specify the number of NFTs fetched when using `alchemy.nft.getNftsForOwner()`.
+- Added the `alchemy.transact.ethGasStationPrice()` convenience method to get the current gas price from ETH Gas Station API.
 
 ## 2.0.2