+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Changelog for Alchemy SDK for JavaScript
 
 ## Unreleased
-- Added support for a Wallet object. This object is identical to the Ethers `Wallet` class.
 
 ### Minor Changes
+
 - Added support for the `pageSize` parameter in `GetNftsForOwnerOptions` to specify the number of NFTs fetched when using `alchemy.nft.getNftsForOwner()`.
 - Added support for a Wallet object. This object is identical to the Ethers `Wallet` class.